requirements.txt
discord.py>=2.3.0
pycountry>=23.12.11
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiopg>=1.4.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
httpx>=0.25.0